Don't let the run/pause toolbar button know when it moves around in the window.
If the titlebar is visible, NSThemeFrame winds up moving toolbar buttons to a new view within the same window. IDERunPauseContinueToolbarButton seems to have a bug when it gets -viewWillMoveToWindow: with the same window as it's already inside of. This method calls down to a helper that unregisters existing KVO notifications and then registers new ones; I can't see why this pattern is failing, but it's a scenario that titlebar-less Xcode doesn't hit, so I'm just going to squelch the -viewWillMoveToWindow: notification.

Addresses issue #1.